requirements.txt
# Audio-DeSilencer Dead Air Detection Requirements

# Core audio processing
pydub>=0.25.1

# Database connectivity
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Web server
Flask>=2.3.0
flask-cors>=4.0.0

# Audio analysis
numpy>=1.24.0
scipy>=1.10.0

# CLI support
argparse>=1.4.0